refactor: extract utilities and clean architecture
