+++ src/domain/block.ts
@@ -1,0 +1,4 @@
+export interface Block {
+  pos: [number, number, number];
+  color: [number, number, number, number];
+}
